+++ src/services/langchain/embedding-service.ts
@@ -1,0 +1,113 @@
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.join(path.dirname(__filename), "vectors", "hsnw");
+
+import path from "path";
+import { fileURLToPath } from "url";
+import { HNSWLib, VectorStore } from "langchain/vectorstores";
+import { RecursiveCharacterTextSplitter } from "langchain/text_splitter";
+import { Embeddings, OpenAIEmbeddings } from "langchain/embeddings";
+
+interface EmbeddingInputBase {
+  embedding?: Embeddings;
+  debug?: boolean;
+}
+
+interface EmbeddingLoadInput extends EmbeddingInputBase {
+  path: string;
+}
+
+interface EmbeddingInput extends EmbeddingInputBase {
+  documents: string[];
+  path?: string;
+  chunkSize?: number;
+}
+
+// Singleton class named EmbeddingDirectory
+class EmbeddingDirectory {
+  private static instance: EmbeddingDirectory;
+
+  private constructor() {}
+
+  public static getInstance(): EmbeddingDirectory {
+    if (!EmbeddingDirectory.instance) {
+      EmbeddingDirectory.instance = new EmbeddingDirectory();
+    }
+
+    return EmbeddingDirectory.instance;
+  }
+
+  public url(url: string): string {
+    // extract hostname and path from url
+    const urlObj = new URL(url);
+    const hostname = urlObj.hostname.replace(/\./g, "-");
+    const urlPath = urlObj.pathname.substring(1);
+
+    return path.resolve(__dirname, "urls", hostname, urlPath);
+  }
+}
+
+class EmbeddingService {
+  public static get embeddingDirectory(): EmbeddingDirectory {
+    return EmbeddingDirectory.getInstance();
+  }
+
+  public static async load(
+    input: EmbeddingLoadInput
+  ): Promise<VectorStore | null> {
+    const debug = input.debug ?? false;
+    const embeddingModel: Embeddings =
+      input.embedding ?? new OpenAIEmbeddings();
+    let vectorStore: HNSWLib;
+
+    // Attempt to load in the vector store
+    if (input.path) {
+      try {
+        if (debug) {
+          console.log(`Loading vector store from path: [${input.path}]`);
+        }
+        // Load in the vector store
+        vectorStore = await HNSWLib.load(input.path, embeddingModel);
+
+        // Return the vector store if it was loaded
+        if (vectorStore) {
+          return vectorStore;
+        }
+      } catch (e) {
+        if (debug) {
+          console.log(`Failed to load vector store from path: [${input.path}]`);
+        }
+      }
+    }
+
+    return null;
+  }
+
+  public static async embed(input: EmbeddingInput): Promise<VectorStore> {
+    const debug = input.debug ?? false;
+    const embeddingModel: Embeddings =
+      input.embedding ?? new OpenAIEmbeddings();
+    let vectorStore: HNSWLib;
+
+    // Create new vector store
+    /* Split the text into chunks */
+    const textSplitter = new RecursiveCharacterTextSplitter({
+      chunkSize: input.chunkSize ?? 1000,
+    });
+    const docs = await textSplitter.createDocuments(input.documents);
+
+    /* Create the vectorstore */
+    vectorStore = await HNSWLib.fromDocuments(docs, embeddingModel);
+
+    // If the path is provided, save the vector store
+    if (input.path) {
+      if (debug) {
+        console.log(`Saving vector store to path: [${input.path}]`);
+      }
+      await vectorStore.save(input.path);
+    }
+
+    return vectorStore;
+  }
+}
+
+export default EmbeddingService;